app: collect status/exception related functionality/classes
